🎨  연관검색엉 없으면 드롭다운 비활성화
